A new file structure has been created for administering users, documents, and news.
