+++ src/app/admin/layout.jsx
@@ -1,0 +1,21 @@
+import Link from "next/link";
+import { getUserRole } from "@/utils/auth/getUserRole";
+
+export default async function AdminLayout({ children }) {
+  const role = await getUserRole();
+
+  return (
+    <div className="w-dvw mx-auto p-6 bg-white/10 rounded shadow-md">
+      <nav className="w-full h-12 flex gap-4 mb-6 text-fuchsia-200">
+        {(role === "moderator" || role === "admin") && (
+          <>
+            <Link href="/admin/documents">Документи</Link>
+            <Link href="/admin/news">Новини</Link>
+          </>
+        )}
+        {role === "admin" && <Link href="/admin/users">Користувачі</Link>}
+      </nav>
+      {children}
+    </div>
+  );
+}
