+++ src/app/admin/news/page.jsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function News() {
+  return <div>News</div>;
+}
