+++ src/app/admin/users/page.jsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function Users() {
+  return <div>Users</div>;
+}
