+++ src/middleware.js
@@ -7,12 +7,12 @@ export default async function middleware(req) {
   const token = await getToken({ req, secret: process.env.NEXTAUTH_SECRET });
   const { pathname } = req.nextUrl;
 
-  console.log("🚀 ~ Middleware token:", token);
+  // console.log("🚀 ~ Middleware token:", token);
 
   const isProtectedRoute = protectedRoutes.some((route) =>
     pathname.startsWith(route),
   );
-  console.log("🚀 ~ token.role:", token?.role);
+  // console.log("🚀 ~ token.role:", token?.role);
 
   if (isProtectedRoute && !token) {
     return NextResponse.redirect(new URL("/login", req.url));
